requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cryptography==44.0.0
Flask==3.0.2
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
gcloud==0.18.3
googleapis-common-protos==1.67.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.6
importlib_metadata==8.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jws==0.1.3
lxml==5.2.1
MarkupSafe==2.1.5
mistune==3.0.2
mysqlclient==2.2.7
oauth2client==4.1.3
packaging==24.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyMySQL==1.1.1
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.2
zipp==3.21.0
